pyproject.toml
[project]
name = "guild-data-app"
version = "0.1.0"
description = "Project to gather, process and display guild data"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "flake8>=7.3.0",
    "gspread>=6.2.1",
    "mypy>=1.18.2",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "psycopg-binary>=3.2.10",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
